Добавлены права доступа для файлов, добавлены контроллеры для работы с md парсером, с документами, с выдачей прав на документы.

+++ backend/MD.WebAPI/Controllers/DocumentController.cs
@@ -7,6 +7,7 @@ using System.Text;
 using System;
 using MD.Infrastructure;
 using MD.Application;
+using Ardalis.Result;
 
 namespace MD.WebAPI;
 
@@ -16,10 +17,17 @@ namespace MD.WebAPI;
 public class DocumentsController : ControllerBase
 {
     private readonly IDocumentService _documentService;
+    private readonly IPermissionService _permissionService;
+    private readonly ICurrentUser _currentUser;
 
-    public DocumentsController(IDocumentService documentService)
+    public DocumentsController(
+        IDocumentService documentService,
+        IPermissionService permissionService,
+        ICurrentUser currentUser)
     {
         _documentService = documentService;
+        _permissionService = permissionService;
+        _currentUser = currentUser;
     }
 
     [HttpPost("upload")]
@@ -30,12 +38,21 @@ public class DocumentsController : ControllerBase
         if (!uploadingResult.IsSuccess)
             return BadRequest(string.Join(", ", uploadingResult.Errors));
 
+        await _permissionService.GrantAccessAsync(
+            uploadingResult.Value,
+            _currentUser.Id,
+            AccessLevel.Edit
+        );
+
         return Ok(new { DocumentId = uploadingResult.Value});
     }
 
     [HttpGet("download/{id}")]
     public async Task<IActionResult> DownloadDocument(Guid id)
     {
+        if (!await _permissionService.CheckAccessAsync(id, _currentUser.Id, AccessLevel.Read))
+            return Forbid();
+
         var downloadingResult = await _documentService.DownloadDocumentAsync(id);
 
         if (!downloadingResult.IsSuccess)
@@ -49,6 +66,9 @@ public class DocumentsController : ControllerBase
     [HttpDelete("delete/{id}")]
     public async Task<IActionResult> DeleteDocument(Guid id)
     {
+        if (!await _permissionService.CheckAccessAsync(id, _currentUser.Id, AccessLevel.Read))
+            return Forbid();
+
         var deletingResult = await _documentService.DeleteDocumentAsync(id);
 
         if (!deletingResult.IsSuccess)
@@ -57,3 +77,5 @@ public class DocumentsController : ControllerBase
         return NoContent();
     }
 }
+
+public record ShareRequest(Guid UserId, AccessLevel AccessLevel);
